Replaced boxart for Mega Man Zero 4 (Europe)
In my previous PR I said that "Mega Man Zero 4 (Europe)" has a small tear on the bottom.
I managed to fix it in GIMP. I asked if it's okay to upload it instead, but got no answer. So I'm asking again.
Also RobLoach seems to have merged my PR as I added to it new covers for the Spyro games. (Not sure if he intended to merge those too.)
mega man zero 4 europe e