Replaced boxart for Yggdra Union EU
Replaced boxart for "Yggdra Union - We'll Never Fight Alone (Europe)". It previously had the USA boxart, but this is comparable in quality. It has some bright lines where the sun reflects of it (I replaced the gameboy advance logo on the left since it also had bright lines on it, and it's less distracting like this).